requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==3.0.10
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
mariadb==1.1.11
MarkupSafe==3.0.2
numpy==2.2.2
ollama==0.4.7
packaging==24.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
pyngrok==7.2.3
python-dateutil==2.9.0.post0
python-ngrok==0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
